Crate rascam

Access the native Raspberry Pi camera.

This module uses the MMAL library (mmal-sys) to access the Raspberry Pi's camera in a friendly way.

Structs

BufferGuard
Guard around a buffer header.

CameraError
Represents any error returned when calling a camera function.

CameraInfo
Information about an attached camera. Created by the info function.

CameraSettings
Settings for the camera.

Info
Contains information about attached cameras.

MmalError
Represents an error from the MMAL library.

SeriousCamera

SimpleCamera
A simple camera interface for the Raspberry Pi

Constants

ISO_100

ISO_125

ISO_160

ISO_200

ISO_250

ISO_320

ISO_400

ISO_500

ISO_640

ISO_800

ISO_1000

ISO_1250

ISO_1600

ISO_2000

ISO_2500

ISO_3200

ISO_AUTO

MMAL_ENCODING_GIF

MMAL_ENCODING_JPEG

MMAL_ENCODING_OPAQUE
VideoCore opaque image format, image handles are returned to the host but not the actual image data.

MMAL_ENCODING_PNG

MMAL_ENCODING_RGB24

Functions

drop_port_userdata
Drops a port's userdata. userdata must be non-null or will dereference a null pointer!

info
Retrieves info on attached cameras

Type Definitions

ISO